[refactor] exception 경로 변경
