+++ src/main/java/com/hhplus/concert/api/token/application/TokenService.java
@@ -3,9 +3,9 @@ package com.hhplus.concert.api.token.application;
 import com.hhplus.concert.api.token.domain.repository.QueueTokenRepository;
 import com.hhplus.concert.api.token.domain.entity.QueueToken;
 import com.hhplus.concert.api.token.domain.type.TokenStatus;
-import com.hhplus.concert.exception.list.CustomBadRequestException;
-import com.hhplus.concert.exception.list.CustomForbiddenException;
-import com.hhplus.concert.exception.list.CustomNotFoundException;
+import com.hhplus.concert.common.exception.list.CustomBadRequestException;
+import com.hhplus.concert.common.exception.list.CustomForbiddenException;
+import com.hhplus.concert.common.exception.list.CustomNotFoundException;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
