+++ src/main/java/com/hhplus/concert/common/exception/list/CustomBadRequestException.java
@@ -1,4 +1,4 @@
-package com.hhplus.concert.exception.list;
+package com.hhplus.concert.common.exception.list;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -6,11 +6,10 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @RequiredArgsConstructor
-public class CustomNotFoundException extends RuntimeException {
-
+public class CustomBadRequestException extends RuntimeException{
     private HttpStatus errorCode;
 
-    public CustomNotFoundException(HttpStatus errorCode, String message) {
+    public CustomBadRequestException(HttpStatus errorCode, String message) {
         super(message);
         this.errorCode = errorCode;
     }
