+++ src/main/java/com/hhplus/concert/common/exception/list/CustomNotFoundException.java
@@ -1,4 +1,4 @@
-package com.hhplus.concert.exception.list;
+package com.hhplus.concert.common.exception.list;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -6,10 +6,11 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @RequiredArgsConstructor
-public class CustomBadRequestException extends RuntimeException{
+public class CustomNotFoundException extends RuntimeException {
+
     private HttpStatus errorCode;
 
-    public CustomBadRequestException(HttpStatus errorCode, String message) {
+    public CustomNotFoundException(HttpStatus errorCode, String message) {
         super(message);
         this.errorCode = errorCode;
     }
